cssom-view/elementFromPoint.html

<!DOCTYPE html>
<title>cssom-view - ElementFromPoint</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  .size {
    width:100px;
    height:100px;
  }
  .overlay {
    position:absolute;
    top:109px;
    pointer-events:none;
  }
  .purple {
    background-color: rebeccapurple;
  }
  .red {
    background-color: red;
  }
  .green {
    background-color: green;
  }
</style>
<body>
  <div id='purple' class="size purple" >&nbsp;</div>
  <div id='red' class="size red">&nbsp;</div>
  <div id='green' class="size overlay green">&nbsp;</div>
  <script>
    assert_equals(null, document.elementFromPoint(-1, -1),
                  "If co-ords are negative should return null");
    assert_equals(document.getElementById('purple'), document.elementFromPoint(10, 10),
                  "Return the first element that is the target for hit testing");
    assert_equals(document.getElementById('red'), document.elementFromPoint(10, 120),
                  "should return the first element that gets the mouse event with pointer-events css");
    done()
 </script>
